+++ api/sample.py
@@ -13,10 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from typing import List
 
-from fastapi import APIRouter, Depends, status
-from sqlalchemy import select
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 from starlette.status import HTTP_201_CREATED
 
@@ -27,13 +25,8 @@ from db.engine import get_db_session
 from db.sample.geothermal import GeothermalSample
 from db.sample.geochemical import GeochemicalSample
 from db.sample.sample import Sample
-from schemas.create.sample import (
-    CreateSample,
-    CreateGeochemicalSample,
-    CreateGeothermalSample,
-)
-from schemas.response.sample import SampleResponse
-from services.query_helper import simple_all_getter, paginated_all_getter
+from schemas_v2.sample import SampleResponse, CreateSample, CreateGeochemicalSample, CreateGeothermalSample
+from services.query_helper import paginated_all_getter
 
 router = APIRouter(
     prefix="/sample",
