--- schemas/create/chemistry.py
@@ -1,36 +1,0 @@
-# ===============================================================================
-# Copyright 2025 ross
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ===============================================================================
-from datetime import datetime
-
-from pydantic import BaseModel
-
-
-# class CreateAnalysisSet(BaseModel):
-#     well_id: int  # mapped_column(Integer, ForeignKey('well.id'))
-#     laboratory: str | None = None  # mapped_column(String(100), nullable=True)
-#
-#     collection_timestamp: datetime
-#
-#
-# class CreateWaterChemistryAnalysis(BaseModel):
-#     analysis_set_id: int
-#     value: float
-#     unit: str
-#     analyte: str
-#     analysis_timestamp: datetime | None = None
-
-
-# ============= EOF =============================================
