--- schemas/response/publication.py
@@ -1,44 +1,0 @@
-# ===============================================================================
-# Copyright 2025 ross
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ===============================================================================
-from pydantic import BaseModel
-
-
-class AuthorResponse(BaseModel):
-    """
-    Schema for the response of an author.
-    """
-
-    id: int
-    name: str
-    email: str | None = None
-    affiliation: str | None = None
-
-
-class PublicationResponse(BaseModel):
-    """
-    Schema for the response of a publication.
-    """
-
-    id: int
-    title: str
-    authors: list[AuthorResponse]
-    year: int
-    doi: str | None = None
-    url: str | None = None
-    publication_type: str
-
-
-# ============= EOF =============================================
