+++ schemas_v2/sample.py
@@ -18,6 +18,11 @@ from datetime import datetime
 from pydantic import BaseModel
 
 
+class SampleResponse(BaseModel):
+    id: int
+
+
+# ============= EOF =============================================
 class CreateSample(BaseModel):
     collection_timestamp: datetime
     collection_method: str
@@ -38,16 +43,3 @@ class CreateGeothermalSample(BaseModel):
     """
 
     sample_id: int
-
-
-# ============= EOF =============================================
-# class CreateWellTimeseries(BaseModel):
-#     name: str
-#     description: str | None = None
-#     well_id: int
-#
-#
-# class CreateGroundwaterLevelObservation(BaseModel):
-#     timestamp: datetime  # ISO 8601 format
-#     value: float
-#     timeseries_id: int
